fix(c-api) Fix memory leak in `WasmFunctionType`.
`WasmFunctionType` has a `params` and a `results` fields. Those fields
have the `Box<wasm_valtype_vec_t>` type. This `wasm_valtype_vec_t` has
no `Drop` implementation, so `Box<wasm_valtype_vec_t>` does not drop,
which is the source of the leak.

We can solve this problem by implementing `Drop` for
`wasm_valtype_vec_t`, in the `macros` module for example. That would
be very convenient. Then `Box<wasm_valtype_vec_t>` would drop
properly, and the memory leak would be solved. Neat!

But we have to deal with another pattern in our code. Some functions
use the `Box<T>` pattern on arguments to express that the function
takes ownerships of the values, like in `wasm_functype_new`. This
pattern is actually an error in this case if the `Box<T>` is
dropped. We must drop the value `T` but not the `Box` itself,
otherwise we might have a double-free in circumstances (if the value
passed by reference to the function is allocated on the stack, C will
free it, which will create a double-free).

So the `Box<wasm_valtype_vec_t>` must not drop. We must free its value
manually. The best way to do so is to call the
`wasm_valtype_vec_delete` function manually in a new `Drop`
implementation for the `WasmFunctionType` type, and inside
`wasm_functype_new`. In both situations, one will need to write:

```rust
wasm_valtype_vec_delete(Some(params.as_mut()));
wasm_valtype_vec_delete(Some(params.as_mut()));
```

In addition to:

```rust
mem::forget(params);
mem::forget(results);
```

inside the `wasm_functype_new` function.

That's a lot of repetitive code, and mostly error-prone.

That's why this PR proposes to solve the problem by using a `Own`
type. This new `Own` type is like a `Box`: it has the same size has a
`Box`, it derefs to `Box` and has a transparent representation to the
inner `Box`; except that it has 2 ways to destruct it:

* By using the `Drop` implementation,
* By calling `Own::drop_value`.

So, for types that own a `Box`, one need to change to `Own` and boom,
no need to write anything more. The `Drop` implementation will take
care of cleaning up the value, and the pointer.

And for functions that take ownerships of the received pointer, one
need to call `Own::drop_value` to drop value without dropping the
pointer.